consolidate shared harness code in harnesses.utils
The building blocks that standalone chat programs share were spread
over verifiers.v1.mcp.client, harnesses/standalone.py, and a monolithic
minimal/program.py. Gather them in verifiers.v1.harnesses.utils:

- mcp.py: the bundleable MCP client helpers (from mcp/client.py)
- compaction.py: tool-output bounding and the Compactor (from
  minimal/program.py)
- core.py: the chat call, local tools, interception hook, and chat
  loop (from minimal/program.py)
- launch.py: launch_chat_program and bundle_program, which generalizes
  inline_mcp_client to splice any utils modules into a PEP 723 program
  (from standalone.py)

minimal/program.py shrinks to argument parsing and assembly. The
verifiers.v1.mcp package keeps the host-side server infrastructure and
no longer re-exports the client helpers (re-exporting would import the
whole harnesses package during mcp's own import). Drops a duplicated
CONTEXT_OVERFLOW_MARKERS tuple left by a merge, and narrows the
.gitignore core-dump pattern to core.[0-9]* so it stops swallowing
core.py.
