+++ verifiers/v1/harnesses/null/harness.py
@@ -2,7 +2,7 @@ from verifiers.v1.clients import ModelContext
 from verifiers.v1.configs.harness import HarnessConfig
 from verifiers.v1.harness import Harness
 from verifiers.v1.harnesses.minimal import PROGRAM_SOURCE
-from verifiers.v1.harnesses.standalone import launch_chat_program
+from verifiers.v1.harnesses.utils.launch import launch_chat_program
 from verifiers.v1.runtimes import ProgramResult, Runtime
 from verifiers.v1.task import TaskData
 from verifiers.v1.trace import Trace
