+++ verifiers/v1/harnesses/utils/launch.py
@@ -1,23 +1,23 @@
 import inspect
 import json
 from collections.abc import Sequence
+from types import ModuleType
 
 from verifiers.v1.clients import ModelContext
 from verifiers.v1.configs.harness import HarnessConfig
 from verifiers.v1.dialects.chat import message_to_wire
-from verifiers.v1.mcp import client as mcp_client
 from verifiers.v1.runtimes import ProgramResult, Runtime
 from verifiers.v1.trace import Trace
 from verifiers.v1.types import Messages
 
-MCP_CLIENT_SOURCE = inspect.getsource(mcp_client)
 PEP_723_END = "# ///\n"
 
 
-def inline_mcp_client(program: str) -> str:
-    """Embed the public client so PEP 723 programs need only their declared packages."""
+def bundle_program(program: str, *modules: ModuleType) -> str:
+    """Embed utils modules so PEP 723 programs need only their declared packages."""
     metadata, body = program.split(PEP_723_END, 1)
-    return f"{metadata}{PEP_723_END}{MCP_CLIENT_SOURCE}\n{body}"
+    sources = "\n".join(inspect.getsource(module) for module in modules)
+    return f"{metadata}{PEP_723_END}{sources}\n{body}"
 
 
 async def launch_chat_program(
