+++ verifiers/v1/mcp/__init__.py
@@ -1,10 +1,3 @@
-from verifiers.v1.mcp.client import (
-    call_mcp,
-    connect_mcp,
-    mcp_client,
-    mcp_content_to_chat_content,
-    with_retry,
-)
 from verifiers.v1.mcp.launch import (
     SharedToolServer,
     serve,
@@ -20,12 +13,7 @@ __all__ = [
     "SharedToolsetConfig",
     "Toolset",
     "ToolsetConfig",
-    "call_mcp",
-    "connect_mcp",
-    "mcp_client",
-    "mcp_content_to_chat_content",
     "serve",
     "serve_shared",
     "serve_tools",
-    "with_retry",
 ]
